Move api mixin to lib, remove mixin
The mixin doesn't gain us anything except complexity
